updates based on feedback
