+++ pkg/combustion/templates/48-message.sh.tpl
